Original XBONE UI mod
There are three versions of the mod: Standard Layout, Western Layout, and PE layout.

Standard is the same as the original game, where the action button is the rightmost (B) face button.
Western is more similar to other western games where the action button is the bottom (A) button. 
PE is the same as Western except the top (Y) button is to attack and the left (X) button is to jump.